[24/10/13] 1B019
- Entity 'moving' reworked, now double value as speed is allowed

+++ src/com/mar21/au/Game.java
@@ -128,9 +128,9 @@ public class Game extends Canvas implements Runnable {
 		}
 
 		screen.clear();
-		int xScroll = player.getX() - screen.width / 2;
-		int yScroll = player.getY() - screen.height / 2;
-		level.render(xScroll, yScroll, screen);
+		double xScroll = player.getX() - screen.width / 2;
+		double yScroll = player.getY() - screen.height / 2;
+		level.render((int) xScroll, (int) yScroll, screen);
 
 		for (int i = 0; i < pixels.length; i++) {
 			pixels[i] = screen.pixels[i];
