+++ src/com/mar21/au/entity/Entity.java
@@ -8,7 +8,7 @@ import com.mar21.au.level.Level;
 
 public class Entity {
 
-	protected int x, y;
+	protected double x, y;
 	protected Sprite sprite;
 	private boolean removed = false;
 	protected Level level;
@@ -37,7 +37,7 @@ public class Entity {
 	}
 
 	public void render(Screen screen) {
-		if (sprite != null) screen.renderSprite(x, y, sprite, true);
+		if (sprite != null) screen.renderSprite((int) x, (int) y, sprite, true);
 	}
 
 	public void remove() {
@@ -52,11 +52,11 @@ public class Entity {
 		this.level = level;
 	}
 
-	public int getX() {
+	public double getX() {
 		return x;
 	}
 
-	public int getY() {
+	public double getY() {
 		return y;
 	}
